version1.3 OLED显示状态

+++ Application/pid_task.c
@@ -18,12 +18,13 @@ void PID_Task(void *argument){
         angle_speed = FULL_PID_Realize(&pid_angle, Target_Angle,angle_Car_total);
 
         //输出PID计算结果
-        motorA_run(pid_speed_A.output);
-        motorB_run(pid_speed_B.output);
-        motorC_run(pid_speed_C.output);
-        motorD_run(pid_speed_D.output);
+        motorA_run((int)pid_speed_A.output);
+        motorB_run((int)pid_speed_B.output);
+        motorC_run((int)pid_speed_C.output);
+        motorD_run((int)pid_speed_D.output);
 
         vTaskDelayUntil(&CurrentTime_PID,5);
+
     }
 }
 //void PID_Timer_Callback(TimerHandle_t pxTimer ) {
